Fixed type

+++ hestia/models.go
@@ -230,7 +230,7 @@ type Voucher struct {
 type VoucherV2 struct {
 	Id            string           `firestore:"id" json:"id"`
 	CreatedTime   int64            `firestore:"created_time" json:"created_time"`
-	AmountEuro    float64          `firestore:"amount_euro" json:"amount_euro"`
+	AmountEuro    int64          `firestore:"amount_euro" json:"amount_euro"`
 	UserPayment   Payment          `firestore:"user_payment" json:"user_payment"`
 	Status        VoucherStatusV2  `firestore:"status" json:"status"`
 	RefundAddress string           `firestore:"refund_address" json:"refund_address"`
